+++ app/src/errors/tooManyBuckets.error.js
@@ -1,0 +1,10 @@
+class TooManyBucketsError extends Error {
+
+    constructor(status, message) {
+        super(message);
+        this.status = status;
+    }
+
+}
+
+module.exports = TooManyBucketsError;